refactor(backoffice-v2): moved every block from useTasks to a hook of its own
