+++ apps/backoffice-v2/package.json
@@ -50,7 +50,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@ballerine/blocks": "0.1.23",
     "@ballerine/common": "0.7.30",
     "@ballerine/ui": "0.3.21",
     "@ballerine/workflow-browser-sdk": "0.5.30",
