+++ apps/backoffice-v2/src/pages/Entity/Entity.page.tsx
@@ -1,9 +1,11 @@
 import { Case } from './components/Case/Case';
 import { useEntity } from './hooks/useEntity/useEntity';
+import { ctw } from '../../common/utils/ctw/ctw';
+import { Card } from '../../common/components/atoms/Card/Card';
+import { CardContent } from '../../common/components/atoms/Card/Card.Content';
 import { KycBlock } from './components/KycBlock/KycBlock';
 import { NoTasksSvg } from '../../common/components/atoms/icons';
 import { ChildDocumentBlocks } from '@/pages/Entity/components/ChildDocumentBlocks/ChildDocumentBlocks';
-import { BlocksComponent } from '@ballerine/blocks';
 
 export const Entity = () => {
   const {
@@ -31,9 +33,33 @@ export const Entity = () => {
         }
       />
       <Case.Content key={selectedEntity?.id}>
-        <BlocksComponent blocks={tasks} cells={cells}>
-          {(Cell, cell) => <Cell {...cell} />}
-        </BlocksComponent>
+        {Array.isArray(tasks) &&
+          tasks?.length > 0 &&
+          tasks?.map((task, index) => {
+            if (!Array.isArray(task?.cells) || !task?.cells?.length) return;
+
+            return (
+              <Card
+                key={index}
+                className={ctw(
+                  'me-4 shadow-[0_4px_4px_0_rgba(174,174,174,0.0625)]',
+                  task.className,
+                )}
+              >
+                <CardContent
+                  className={ctw('grid gap-2', {
+                    'grid-cols-2': task?.cells.some(field => field?.type === 'multiDocuments'),
+                  })}
+                >
+                  {task?.cells.map((field, index) => {
+                    const Cell = cells[field?.type];
+
+                    return <Cell key={index} {...field} />;
+                  })}
+                </CardContent>
+              </Card>
+            );
+          })}
         {kybChildWorkflows?.map(childWorkflow => (
           <ChildDocumentBlocks
             parentWorkflowId={workflow?.id}
