+++ apps/backoffice-v2/src/pages/Entity/components/Heading/Heading.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from 'react';
+import { IHeadingProps } from './interfaces';
 import { ctw } from '../../../../common/utils/ctw/ctw';
-import { ExtractCellProps } from '@ballerine/blocks';
 
-export const Heading: FunctionComponent<ExtractCellProps<'heading'>> = ({ id, value, props }) => (
+export const Heading: FunctionComponent<IHeadingProps> = ({ id, value, props }) => (
   <h2
     className={ctw(
       `ml-1 mt-6 px-2 text-2xl font-bold`,
