+++ apps/backoffice-v2/src/pages/Entity/components/Heading/interfaces.ts
@@ -1,0 +1,7 @@
+import { ComponentProps } from 'react';
+
+export interface IHeadingProps {
+  id?: string;
+  value: string;
+  props?: ComponentProps<'h2'>;
+}
