+++ apps/backoffice-v2/src/pages/Entity/components/KycBlock/KycBlock.tsx
@@ -4,8 +4,7 @@ import { useKycBlock } from './hooks/useKycBlock/useKycBlock';
 import { Card } from '../../../../common/components/atoms/Card/Card';
 import { CardContent } from '../../../../common/components/atoms/Card/Card.Content';
 import { ctw } from '../../../../common/utils/ctw/ctw';
-
-import { cells } from '@/pages/Entity/hooks/typed-create-blocks/typed-create-blocks';
+import { cells } from '@/pages/Entity/hooks/useEntity/cells';
 
 export const KycBlock: FunctionComponent<{
   parentWorkflowId: string;
