+++ apps/backoffice-v2/src/pages/Entity/hooks/useEntity/useEntity.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router-dom';
 import { useFilterId } from '../../../../common/hooks/useFilterId/useFilterId';
 import { useWorkflowQuery } from '../../../../domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
 import { useTasks } from '../useTasks/useTasks';
-import { cells } from '@/pages/Entity/hooks/typed-create-blocks/typed-create-blocks';
+import { cells } from '@/pages/Entity/hooks/useEntity/cells';
 
 export const useEntity = () => {
   const { entityId } = useParams();
