+++ apps/backoffice-v2/src/pages/Entity/hooks/useTasks/hooks/useDirectorsRegistryProvidedBlock/useDirectorsRegistryProvidedBlock.tsx
@@ -1,0 +1,39 @@
+export const useDirectorsRegistryProvidedBlock = directorsRegistryProvided => {
+  if (!Array.isArray(directorsRegistryProvided) || !directorsRegistryProvided?.length) {
+    return [];
+  }
+
+  return [
+    {
+      cells: [
+        {
+          type: 'heading',
+          value: 'Directors',
+        },
+        {
+          type: 'subheading',
+          value: 'Registry-provided Data',
+          props: {
+            className: 'mb-4',
+          },
+        },
+        {
+          type: 'table',
+          value: {
+            columns: [
+              {
+                accessorKey: 'name',
+                header: 'Name',
+              },
+              {
+                accessorKey: 'position',
+                header: 'Position',
+              },
+            ],
+            data: directorsRegistryProvided,
+          },
+        },
+      ],
+    },
+  ];
+};
